refactor: change directory of login form, and fix typo on font name (#3)
