+++ src/assets/less/font.less
@@ -1,26 +1,26 @@
 @font-face {
-  font-family: "LouisGoergeCafe";
+  font-family: "LouisGeorgeCafe";
   src: url("../fonts/Louis-George-Cafe/Louis-George-Cafe-Italic.ttf");
   font-weight: 400;
   font-style: italic, oblique;
 }
 
 @font-face {
-  font-family: "LouisGoergeCafe";
+  font-family: "LouisGeorgeCafe";
   src: url("../fonts/Louis-George-Cafe/Louis-George-Cafe-Light.ttf");
   font-weight: 300;
   font-style: normal;
 }
 
 @font-face {
-  font-family: "LouisGoergeCafe";
+  font-family: "LouisGeorgeCafe";
   src: url("../fonts/Louis-George-Cafe/Louis-George-Cafe-Bold.ttf");
   font-weight: bold;
   font-style: normal;
 }
 
 @font-face {
-  font-family: "LouisGoergeCafe";
+  font-family: "LouisGeorgeCafe";
   src: url("../fonts/Louis-George-Cafe/Louis-George-Cafe.ttf");
   font-weight: 400;
   font-style: normal;
@@ -29,5 +29,5 @@
 *,
 *::before,
 ::after {
-  font-family: "LouisGoergeCafe";
+  font-family: "LouisGeorgeCafe";
 }
